fazendo funcionar
