+++ modulo3/apis-labenusers/aula-api/src/App.js
@@ -1,30 +1,31 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import ListaUsuarios from "./Components/Usuarios/ListaUsuarios";
 import AdicionarUsuario from './Components/Adicionar/AdicionarUsuario';
+import axios from "axios";
 
 function App() {
 
-  // const [tela, setTela] = useState("")
+  const [foiTrocado, setFoiTrocado] = useState(false)
+  const [users, setUsers] = useState([])
 
-  // const trocarTela = () => {
-  //   if (tela === "Criar") {
-  //     return (
-  //       <div className="App">
-  //         <ListaUsuarios setTela={setTela} />
-  //       </div>
-  //     )
-  //   } if (tela === "Usuário") {
-  //     return (
-  //       <div className="App">
-  //         <ListaUsuarios setTela={setTela} />
-  //       </div>
-  //     )
-  //   }
-  // }
+  const userPass = users
+
+  const url = 'https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users';
+  const serviceHeaders = { headers: { Authorization: "juan-pedroza-jemison" } }
+
+  const getAllUsers = () => {
+    axios.get(url, serviceHeaders)
+        .then((response) => setUsers(response.data))
+        .catch((error) => console.log(error.response))
+}
+  useEffect(() => {
+    getAllUsers()
+  }, [])
 
   return (
     <div className="App">
-      <ListaUsuarios />
+      <button onClick={() => setFoiTrocado(!foiTrocado)}>Trocar de tela</button>
+      {foiTrocado ? <ListaUsuarios url={url} serviceHeaders={serviceHeaders} getAllUsers={getAllUsers()} userPass={userPass} /> : <AdicionarUsuario url={url} serviceHeaders={serviceHeaders} />}
     </div>
   );
 }
